Restructured the project and renamed to SuperMart
