--- Assignment.Dal/Extensions/StringExtensions.cs
@@ -1,12 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-
-namespace Assignment.Dal.Extensions
-{
-    public static class StringExtensions
-    {
-        public static DbContextOptions<DbContext> StringToContextOptions(this string connectionString)
-        {
-            return new DbContextOptionsBuilder<DbContext>().UseMySQL(connectionString).Options;
-        }
-    }
-}
