--- Assignment.Dal/Repositories/IProductsRepository.cs
@@ -1,15 +1,0 @@
-﻿using Assignment.Dal.Entities;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace Assignment.Dal.Repositories
-{
-    public interface IProductsRepository
-    {
-        Task AddProductAsync(string storeName, Product product);
-        Task<(IEnumerable<Product>, long)> GetAllStoreProductsAsync(string storeName);
-        Task<Product> GetProductAsync(string storeName, string productName);
-        Task RemoveProductAsync(string storeName, string productName);
-        Task<(IEnumerable<Product>, long)> GetAllProductsAsync(string searchText = null);
-    }
-}
