--- Assignment.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30711.63
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Assignment.Dal", "Assignment.Dal\Assignment.Dal.csproj", "{3CA484FF-9C9B-4DD5-9A96-0E5C01D29289}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Assignment.Apis", "Assignment.Apis\Assignment.Apis.csproj", "{D4AF73F8-95B7-4119-A205-964E26878372}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3CA484FF-9C9B-4DD5-9A96-0E5C01D29289}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3CA484FF-9C9B-4DD5-9A96-0E5C01D29289}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3CA484FF-9C9B-4DD5-9A96-0E5C01D29289}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3CA484FF-9C9B-4DD5-9A96-0E5C01D29289}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D4AF73F8-95B7-4119-A205-964E26878372}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D4AF73F8-95B7-4119-A205-964E26878372}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D4AF73F8-95B7-4119-A205-964E26878372}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D4AF73F8-95B7-4119-A205-964E26878372}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {90015199-90F3-4A27-AFE4-BBE17CBF0F82}
-	EndGlobalSection
-EndGlobal
