+++ SuperMart.Apis/Extensions/ProductExtensions.cs
@@ -1,0 +1,33 @@
+﻿using SuperMart.Apis.Models;
+using System;
+
+namespace SuperMart.Apis.Extensions
+{
+    public static class ProductExtensions
+    {
+        public static Dal.Entities.Product ToDalEntity(this ProductInputModel input)
+        {
+            return new Dal.Entities.Product
+            {
+                Name = input.Name,
+                Description = input.Description,
+                Category = input.Category,
+                Price = input.Price,
+                AddedOn = DateTime.Now
+            };
+        }
+
+        public static ProductOutputModel ToOutputModel(this Dal.Entities.Product input)
+        {
+            return new ProductOutputModel
+            {
+                Name = input.Name,
+                Description = input.Description,
+                Category = input.Category,
+                Price = input.Price,
+                AddedOn = DateTime.Now,
+                StoreName = input.Store?.StoreName
+            };
+        }
+    }
+}
