+++ SuperMart.Apis/Models/ProductInputModel.cs
@@ -1,0 +1,10 @@
+﻿namespace SuperMart.Apis.Models
+{
+    public class ProductInputModel
+    {
+        public string Name { get; set; }
+        public string Description { get; set; }
+        public double Price { get; set; }
+        public string Category { get; set; }
+    }
+}
