+++ SuperMart.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30711.63
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SuperMart.Dal", "SuperMart.Dal\SuperMart.Dal.csproj", "{1B62324A-4E22-4BF2-99DF-314A1F20B3C3}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SuperMart.Apis", "SuperMart.Apis\SuperMart.Apis.csproj", "{6A11C7D0-7461-4196-922C-0E0711314250}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{1B62324A-4E22-4BF2-99DF-314A1F20B3C3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1B62324A-4E22-4BF2-99DF-314A1F20B3C3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1B62324A-4E22-4BF2-99DF-314A1F20B3C3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1B62324A-4E22-4BF2-99DF-314A1F20B3C3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6A11C7D0-7461-4196-922C-0E0711314250}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6A11C7D0-7461-4196-922C-0E0711314250}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6A11C7D0-7461-4196-922C-0E0711314250}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6A11C7D0-7461-4196-922C-0E0711314250}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {90015199-90F3-4A27-AFE4-BBE17CBF0F82}
+	EndGlobalSection
+EndGlobal
